requirements.txt
aws-cdk-lib==2.199.0
constructs>=10.0.0,<11.0.0
pydantic>=2.0.0,<3.0.0
pyyaml>=6.0.0,<7.0.0
